# core/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Deck OS - JARVIS Cyberdeck Command Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai-router/infer:
    post:
      operationId: routeInference
      tags: [ai-router]
      summary: Route an inference request to the best available model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InferenceRequest"
      responses:
        "200":
          description: Inference response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InferenceResponse"
  /ai-router/mode:
    get:
      operationId: getIntelligenceMode
      tags: [ai-router]
      summary: Get current intelligence mode
      responses:
        "200":
          description: Current intelligence mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntelligenceMode"
    put:
      operationId: setIntelligenceMode
      tags: [ai-router]
      summary: Set intelligence mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetIntelligenceModeRequest"
      responses:
        "200":
          description: Updated intelligence mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntelligenceMode"
...
  /commands:
    get:
      operationId: listCommands
      tags: [commands]
      summary: List all available commands
      responses:
        "200":
          description: Available commands
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandList"
    post:
      operationId: dispatchCommand
      tags: [commands]
      summary: Dispatch a command through the skill router
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommandDispatchRequest"
      responses:
        "200":
          description: Command dispatched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
  /commands/history:
    get:
      operationId: getCommandHistory
      tags: [commands]
      summary: Get recent command history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Command history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandHistoryList"
...
  /devices/{deviceId}/command:
    post:
      operationId: sendDeviceCommand
      tags: [devices]
      summary: Send a command to a device via the event bus
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceCommandRequest"
      responses:
        "200":
          description: Command result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceCommandResult"
        "404":
          description: Device not found
...
components:
  schemas:
...
    InferenceRequest:
      type: object
      properties:
        prompt:
          type: string
        mode:
          type: string
          enum: [fast, deep, none]
        pluginId:
          type: string
          nullable: true
        context:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
              content:
                type: string
            required:
              - role
              - content
        useCache:
          type: boolean
          default: true
      required:
        - prompt
        - mode
    InferenceResponse:
      type: object
      properties:
        response:
          type: string
        modelUsed:
          type: string
        mode:
          type: string
        fromCache:
          type: boolean
        latencyMs:
          type: integer
        tokens:
          type: integer
          nullable: true
      required:
        - response
        - modelUsed
        - mode
        - fromCache
        - latencyMs
    IntelligenceMode:
      type: object
      properties:
        mode:
          type: string
          enum: [DIRECT_EXECUTION, LIGHT_REASONING, DEEP_REASONING, HYBRID_MODE]
        description:
          type: string
        preferLocal:
          type: boolean
      required:
        - mode
        - description
        - preferLocal
    SetIntelligenceModeRequest:
      type: object
      properties:
        mode:
          type: string
          enum: [DIRECT_EXECUTION, LIGHT_REASONING, DEEP_REASONING, HYBRID_MODE]
      required:
        - mode
...
    CommandList:
      type: object
      properties:
        commands:
          type: array
          items:
            $ref: "#/components/schemas/CommandDef"
      required:
        - commands
    CommandDef:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        plugin:
          type: string
        syntax:
          type: string
        examples:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - description
        - plugin
        - syntax
        - examples
    CommandDispatchRequest:
      type: object
      properties:
        input:
          type: string
        mode:
          type: string
          enum: [DIRECT_EXECUTION, LIGHT_REASONING, DEEP_REASONING, HYBRID_MODE, auto]
          default: auto
        context:
          type: object
          additionalProperties: true
          nullable: true
      required:
        - input
    CommandResult:
      type: object
      properties:
        success:
          type: boolean
        output:
          type: string
        plugin:
          type: string
          nullable: true
        command:
          type: string
          nullable: true
        executionTimeMs:
          type: integer
        modeUsed:
          type: string
        aiAssisted:
          type: boolean
        timestamp:
          type: string
          format: date-time
      required:
        - success
        - output
        - executionTimeMs
        - modeUsed
        - aiAssisted
        - timestamp
    CommandHistoryList:
      type: object
      properties:
        history:
          type: array
          items:
            $ref: "#/components/schemas/CommandHistoryEntry"
        total:
          type: integer
      required:
        - history
        - total
    CommandHistoryEntry:
      type: object
      properties:
        id:
          type: string
        input:
          type: string
        output:
          type: string
        success:
          type: boolean
        plugin:
          type: string
          nullable: true
        executionTimeMs:
          type: integer
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - input
        - output
        - success
        - executionTimeMs
        - timestamp
...
    DeviceCommandRequest:
      type: object
      properties:
        action:
          type: string
        parameters:
          type: object
          additionalProperties: true
          nullable: true
      required:
        - action
    DeviceCommandResult:
      type: object
      properties:
        success:
          type: boolean
        deviceId:
          type: string
        action:
          type: string
        message:
          type: string
      required:
        - success
        - deviceId
        - action
        - message
...